+++ espflash/src/flasher/mod.rs
@@ -488,6 +488,8 @@ pub struct FlashData {
     pub chip: Chip,
     /// Crystal frequency.
     pub xtal_freq: XtalFrequency,
+    /// Whether to add secure boot V2 padding.
+    pub secure_pad_v2: bool,
 }
 
 impl FlashData {
@@ -498,6 +500,7 @@ impl FlashData {
         mmu_page_size: Option<u32>,
         chip: Chip,
         xtal_freq: XtalFrequency,
+        secure_pad_v2: bool,
     ) -> Self {
         FlashData {
             flash_settings,
@@ -505,6 +508,7 @@ impl FlashData {
             mmu_page_size,
             chip,
             xtal_freq,
+            secure_pad_v2,
         }
     }
 }